+++ CHANGELOG.rst
@@ -117,10 +117,12 @@ In the release the MESMER-X functionality is integrated into the MESMER Codebase
 - Enable `threshold_min_proba` to be `None` in `distrib_cov` (`#598 <https://github.com/MESMER-group/mesmer/pull/598>`_).
 - Also use Nelder-Mead fit in `distrib_cov._minimize` for `option_NelderMead == "best_run"` when Powell fit was not successful (`#600 <https://github.com/MESMER-group/mesmer/pull/600>`_).
 - Return `logpmf` for discrete distributions in `distrib_cov._fg_fun_LL_n()` (`#602 <https://github.com/MESMER-group/mesmer/pull/602>`_).
-- Speed-up MESMER-X by
+- Speed-up MESMER-X
+
   - add method to calculate params of a distribution (`#539 <https://github.com/MESMER-group/mesmer/pull/539>`_)
   - avoiding frozen distributions (`#532 <https://github.com/MESMER-group/mesmer/issues/532>`_)
   - not broadcasting scalars (`#613 <https://github.com/MESMER-group/mesmer/pull/613>`_)
+  - compiling the expression (`#614 <https://github.com/MESMER-group/mesmer/pull/614>`_).
 
 
 Integration of MESMER-M